requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# sqlalchemy==2.0.23
# alembic==1.12.1
# psycopg2-binary==2.9.9
# python-jose==3.3.0
# passlib==1.7.4
# python-multipart==0.0.6
# pydantic==2.5.0
# pydantic-settings==2.1.0
# python-dotenv==1.0.0
# bcrypt==4.1.2
# python-dateutil==2.8.2
# email-validator==2.1.0
# jinja2==3.1.2




fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
email-validator==2.1.0
jinja2==3.1.2
